+++ main.lua
@@ -275,7 +275,7 @@ return function(mod)
 
   -- ------- exports (companion / debug / test surface)
 
-  mod.exports.version = "1.8.0"
+  mod.exports.version = "1.9.0"
   mod.exports.logic = logic
   mod.exports.render = render
   mod.exports.animated = render.animated